% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eodplot.R
\name{findLandmarks}
\alias{findLandmarks}
\title{Plot EOD signal with landmarks}
\usage{
findLandmarks(plotdata)
}
\arguments{
\item{plotdata}{The EOD matrix from getEODMatrix}
}
\description{
Plot EOD signal with landmarks
}
