% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eodplot.R
\name{getStats}
\alias{getStats}
\title{Returns a data frame with stats about the landmarks}
\usage{
getStats(peaks, landmark_table)
}
\arguments{
\item{peaks}{list of peaks from peakFinder}

\item{landmark_table}{List of landmark}
}
\description{
Returns a data frame with stats about the landmarks
}
