% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_finder.R
\name{peakFinder}
\alias{peakFinder}
\title{Find peaks from a TDMS file}
\usage{
peakFinder(filename, channel = "/'Untitled'/'Dev1/ai0'", direction = "none",
  threshold = 5, start = NULL, end = NULL, remove = NULL, verbose = F,
  progressCallback = NULL)
}
\arguments{
\item{filename}{The filename}

\item{channel}{The channel name, default /'Untitled'/'Dev1/ai0' which is just common in our lab}

\item{direction}{Only get positive or negative peaks}

\item{threshold}{Threshold for cutoff}

\item{start}{Start of rec, mutually exclusive to remove}

\item{end}{End of rec}

\item{remove}{Remove N seconds from start and end of recording, mutually exclusive to start/end}

\item{verbose}{Verbose output}

\item{progressCallback}{Callback for progress event update}
}
\description{
Find peaks from a TDMS file
}
