% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eodplot.R
\name{plotAverage}
\alias{plotAverage}
\title{Plot average EOD signal}
\usage{
plotAverage(plotdata, verbose = F)
}
\arguments{
\item{plotdata}{The EOD matrix from getEODMatrix}

\item{verbose}{output debug info}
}
\description{
Plot average EOD signal
}
