% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eodplot.R
\name{plotEod}
\alias{plotEod}
\title{Plot EODs and find EOD statistics}
\usage{
plotEod(filename, peaks, channel, prebaseline = F, postbaseline = F,
  normalize = F, alpha = F, window = 0.005, verbose = F)
}
\arguments{
\item{filename}{The filename}

\item{channel}{The channel name, default /'Untitled'/'Dev1/ai0' which is just common in our lab}

\item{prebaseline}{Subtract baseline pre normalization}

\item{postbaseline}{Subtract baseline post normalization}

\item{normalize}{Normalize data to 0-1}

\item{alpha}{Alpha channel for all EODs plot}

\item{window}{Window size}

\item{verbose}{Set verbose output}
}
\description{
Plot EODs and find EOD statistics
}
