% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eodplot.R
\name{plotLandmarks}
\alias{plotLandmarks}
\title{Plot all EOD signal with landmarks}
\usage{
plotLandmarks(plotdata, landmark_table, verbose = F)
}
\arguments{
\item{plotdata}{The EOD matrix from getEODMatrix}

\item{landmark_table}{The landmark table from findLandmarks}

\item{verbose}{output debug info}
}
\description{
Plot all EOD signal with landmarks
}
