% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmsplot.R
\name{plotTdms}
\alias{plotTdms}
\title{Plot the signal vs time for a TDMS file}
\usage{
plotTdms(filename, channel = "/'Untitled'/'Dev1/ai0'", start = 0,
  end = NULL, peaks = NULL)
}
\arguments{
\item{filename}{The filename}

\item{channel}{The channel name, default /'Untitled'/'Dev1/ai0' which is just common in our lab}

\item{start}{Default 0}

\item{end}{Default end of tdms file data}

\item{peaks}{A data frame of peak data and orientations}
}
\description{
Plot the signal vs time for a TDMS file
}
