% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eodplot.R
\name{plotTotal}
\alias{plotTotal}
\title{Plot all EOD signals}
\usage{
plotTotal(plotdata, alpha = 0.05, verbose = F)
}
\arguments{
\item{plotdata}{The EOD matrix from getEODMatrix}
}
\description{
Plot all EOD signals
}
