% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmormyridlandmarks.R
\name{findmormyridlandmarks}
\alias{findmormyridlandmarks}
\title{Find Landmarks from a Mormyrid EOD Waveform}
\usage{
findmormyridlandmarks(data, baseline_n)
}
\arguments{
\item{data}{Two column waveform dataframe, time and voltage x n pts}

\item{baseline_n}{number of points to use in calculating baseline}
}
\description{
Find Landmarks from a Mormyrid EOD Waveform
}
